+++ src/alfasim_sdk/alfasim_sdk_api/api.h
@@ -605,6 +605,20 @@ DLL_EXPORT int get_liqliq_flow_pattern(
 );
 
 /*!
+    Returns the thickness of the deposition of phase on pipe inner wall.
+
+    @param[in] ctx ALFAsim's plugins context.
+    @param[out] out Thickness of phase deposited on inner wall.
+    @param[in] phase_id Phase index.
+    @param[in] ts_scope A #TimestepScope value.
+    @param[in] size Size of out array of values.
+    @return An #error_code value.
+*/
+DLL_EXPORT int get_deposition_thickness(
+   void* ctx, double** out, int phase_id, enum TimestepScope ts_scope, int* size
+);
+
+/*!
     Retrieves the tracer ID given a tracer reference. A tracer reference may be obtained from the
     user input data (See #get_plugin_input_data_reference API function for an example).